Don't mutate global env (#3100)
