+++ packages/core/core/src/resolveOptions.js
@@ -7,7 +7,7 @@ import type {
 } from '@parcel/types';
 
 import {getRootDir} from '@parcel/utils';
-import loadEnv from './loadEnv';
+import loadDotEnv from './loadDotEnv';
 import path from 'path';
 import TargetResolver from './TargetResolver';
 import {resolveConfig} from '@parcel/utils';
@@ -52,13 +52,9 @@ export default async function resolveOptions(
   let targetResolver = new TargetResolver();
   let targets = await targetResolver.resolve(rootDir, cacheDir, initialOptions);
 
-  if (!initialOptions.env) {
-    await loadEnv(path.join(rootDir, 'index'));
-  }
-
   // $FlowFixMe
   return {
-    env: process.env,
+    env: initialOptions.env ?? (await loadDotEnv(path.join(rootDir, 'index'))),
     ...initialOptions,
     cacheDir,
     entries,
