+++ packages/transformers/js/src/visitors/env.js
@@ -1,12 +1,14 @@
 import * as types from '@babel/types';
 
 export default {
-  MemberExpression(node, asset) {
+  MemberExpression(node, {asset, env}) {
     // Inline environment variables accessed on process.env
     if (types.matchesPattern(node.object, 'process.env')) {
       let key = types.toComputedKey(node);
       if (types.isStringLiteral(key)) {
-        let prop = process.env[key.value];
+        // Try using the value from the passed env (either from new Parcel
+        // options or from dotenv), and fall back to process.env
+        let prop = env[key.value] ?? process.env[key.value];
         if (typeof prop !== 'function') {
           let value = types.valueToNode(prop);
           morph(node, value);
